+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - floating button for open console on android
 - saving floating button position
 - saving last opened console tab
+- auto refresh console layout on phone orientation change
 
 ### Changed
 - removed console.switch-context command. its unnecessary